Extract HttpRequestMatcher to separate file
